refactor(bench+doctor): lift model-validation checks from doctor/checks/ to bench/tiers/
PR 1 of 4 — re-architecting doctor (env-health) vs bench (model-validation);
zero behavior change in this PR. Moves six modules out of doctor/checks/
and into a new bench/tiers/ package so a later PR can slim doctor and
another can wire `bench --tier smoke|speed|harness`.

Moves:
  vllm_mlx/doctor/checks/smoke.py     -> vllm_mlx/bench/tiers/smoke.py
  vllm_mlx/doctor/checks/perf.py      -> vllm_mlx/bench/tiers/perf.py
  vllm_mlx/doctor/checks/benchmark.py -> vllm_mlx/bench/tiers/benchmark.py
  vllm_mlx/doctor/checks/stress.py    -> vllm_mlx/bench/tiers/stress.py
  vllm_mlx/doctor/checks/agent.py     -> vllm_mlx/bench/tiers/agent.py
  vllm_mlx/doctor/checks/api.py       -> vllm_mlx/bench/tiers/api.py

user.py stays in doctor/checks/ — it is the only true env-health probe.

Option A (preferred): doctor/cli.py imports were updated to point at the
new path (`from ..bench.tiers import ...`); no re-export shim files left
behind in doctor/checks/. Cleaner diff for PR #4; no import cycles found.

Side-fix: tests/test_no_mllm_flag.py exemption paths updated for the
three moved files that appear there (perf/api/benchmark).
