+++ vllm_mlx/bench/tiers/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Benchmark tiers — one module per validation category.
+
+Each module exposes ``check_*`` / ``benchmark_*`` functions consumed by
+``vllm_mlx.doctor.cli`` (today) and ``vllm_mlx.bench.cli`` (later PR).
+"""
+
+from . import agent, api, benchmark, perf, smoke, stress
+
+__all__ = ["agent", "api", "benchmark", "perf", "smoke", "stress"]
