+++ vllm_mlx/bench/tiers/api.py
@@ -5,7 +5,13 @@ from __future__ import annotations
 
 import time
 
-from ..runner import REPO_ROOT, CheckResult, Status, python_executable, run_subprocess
+from ...doctor.runner import (
+    REPO_ROOT,
+    CheckResult,
+    Status,
+    python_executable,
+    run_subprocess,
+)
 
 
 def check_smoke_matrix(port: int) -> CheckResult:
